Added Dexscreener API script
Given a list of coin tickers ($MYCOIN) or coin names this Jupyter Notebook uses the Dexsreener API (https://docs.dexscreener.com/api/reference) to get onchain cryptocurrency-data. It filters and orders the coin-data in a way that might provide an initial overview over potential short-term investment opportunities.

As Dexscreener does not have an API endpoint to get a list of available coins (yet!), the temporary solution is to manually copy the coin-list from https://dexscreener.com/{my_chain} to a txt-file (dxs-out). Websrapping did not work for me unfortunately.

I have tested this with the Solana-Blockchain so far, but it should work for any Blockchain featured on dexscreener.com